MicroSpice

MicroSpice is a very simple circuit simulator that simulates an electronic circuit composed a variety of passive components. This circuit can be constructed using a subset of standard Spice netlist syntax.

As an example, here is a simple RLC circuit with in as input and out as output

L1 in  out 100nH
R1 out gnd  50Ω
C1 out gnd 100nF

Currently the only supported components are resistors (R), inductors (L) and capacitors (C). Component values can be specified using standard SI prefixes such as k (kilo = 1e3), M (Mega=1e6), G(giga=1e9), m (milli=1e-3), μ (micro=1e-6), n (nano=1e-9), or p (pico=1e-12). In addition, the non-standard forms u (micro=1e-6) and meg (Mega=1e6) are used for Spice compatibility.

Units including H (Henry), F (Farad) and Ω (ohm) are allowed, but are optional, but the unit must match the type of component.

Component values can also be a reference to a parameter. For instance, we could modify the circuit above to parameterize the value of the inductor and capacitor:

L1 in  out $L
R1 out gnd  50Ω
C1 out gnd $C

More on how to use this later

Running a Simulation

A Netlist describes an abstract circuit. To simulate that circuit, the circuit is turned into a matrix equation consisting of complex admittances that relate the voltages and currents in the circuit. In this conversion, the voltages for any subset of nodes can be forced and currents can be injected into any nodes that are not forced to a particular voltage. The result of the voltages and all of the node-to-node currents are then computed.

In normal practice, the internals of this conversion are hidden from view. All you need to do is provide a Netlist and a list of inputs and outputs and you get a simulation function that takes a frequency, a vector of voltages and returns a vector of output voltages. You can call that simulation function repeatedly for different frequencies or different inputs.

Taking the circuit given above, if we take node in as the input, node out as the output and gnd as the ground reference, we can simulate the voltage transfer function for this circuit with this code:

using MicroSpice
nl = MicroSpice.Netlist("L1 in  out 100nH\nR1 out gnd 50Ω\nC1 out gnd 100nF\n")
s = MicroSpice.solve(nl, [:in, :gnd], ["out"])
decibel(x) = 20 * log10(abs(x))
[decibel(only(s(f, [1, 0]))) for f in [1.4e6, 1.5e6, 1.62e6, 1.8e6]]
# output
4-element Vector{Float64}:
 12.883077832402897
 18.914298710446417
 27.65586909592151
 11.05634871505561

Circuit parameters

If you have component values that are parameter references (such as $L in the example above), the values of the parameters must be supplied when you run the actual simulation function. Even before that, though, when you construct the Netlist, you need to provide the names of the parameters that you will be providing later as a list. This list also specifies the order in which you will provide the values later.

Here is an example with the circuit from before but with parameters for the inductor and capacitor. Note the use of a raw string to avoid needing to quote the $ before the parameter names.

using MicroSpice
nl = MicroSpice.Netlist(raw"""
L1 in  out $L
R1 out gnd 50Ω
C1 out gnd $C
""", ["L", "C"])
decibel(x) = 20 * log10(abs(x))
[ 
decibel(only(MicroSpice.solve(nl, [:in, :gnd], ["out"], params)(f, [1, 0])))
for f in  [1.0e6, 1.2e6, 1.4e6, 1.62e6, 1.8e6, 2.0e6],
params in  [[100e-9, 100e-9], [200e-9, 50e-9], [120e-9, 100e-9]]
]
# output
6×3 Matrix{Float64}:
  4.35992   4.35431   5.57245
  7.29487   7.279     9.94251
 12.8831   12.8055   22.5545
 27.6559   25.2887   12.2341
 11.0563   10.9722    5.42304
  4.73621   4.71178   0.958959

MicroSpice.ComplexCircuit

MicroSpice.InternalExpression

MicroSpice.Link

MicroSpice.Netlist

MicroSpice.Netlist

MicroSpice.coil

MicroSpice.inputImpedance

MicroSpice.solve

MicroSpice.solve

MicroSpice.xc

MicroSpice.xl

MicroSpice.ComplexCircuit — Type

A circuit is defined as a set of nodes connected by admittances. Each node may be set to a particular voltage, or can have a current injected or it can be free-floating. This is modeled as a linear system of equations in terms of the n nodal voltages of which m are externally constrained, and k injected currents. The first n of these equations define the relationship between node voltages and injected currents and are of rank n-1. The next m equations define externally set voltages followed by k equations for injected currents. We assume that imposed voltages and injected currents are on different nodes so this gives a rank of n + m + k - 1 so far. There are an additional n-m-k equations that set all remaining injected currents to zero and a final constraint that sets the sum of all injected currents to zero.

MicroSpice.InternalExpression — Type

An InternalExpression is a sum of a numerical value and references into a value vector that will be provided at simulation time.

MicroSpice.Link — Type

A Link is a composite of pure resistance, inductance and capacitance arranged in parallel.

To simplify combining multiple links, resistance and inductance are retained in inverse form (i.e. 1/R, 1/L).

MicroSpice.Netlist — Type

A netlist is defined as a set of nodes which are connected by ideal resistors, inductors and capacitors. Such a netlist can be resolved into a ComplexCircuit at any desired frequency by replacing the inductors and capacitors by their equivalent complex impedances at that frequency.

Constructs a Netlist from a Spice circuit. As an example, here is a simple RLC circuit with in as input and out as output

L1 in  out 100nH
R1 out gnd  50Ω
C1 out gnd 100nF

Currently the only supported components are resistors (R), inductors (L) and capacitors (C). Component values can be specified using standard SI prefixes such as k (kilo = 1e3), M (Mega=1e6), G(giga=1e9), m (milli=1e-3), μ (micro=1e-6), n (nano=1e-9), or p (pico=1e-12). In addition, the non-standard forms u (micro=1e-6) and meg (Mega=1e6) are used for Spice compatibility.

Units including H (Henry), F (Farad) and Ω (ohm) are allowed, but are optional, but only the correct unit type is allowed.f

A Netlist can be converted to a ComplexCircuit at a specific frequency and then solved with specified input voltages and currents to find the voltages for all nodes in the circuit. Taking the example above, we can find the response for a few frequencies using this code

nl = MicroSpice.Netlist("L1 in  out 100n
R1 out gnd 50
C1 out gnd 100n
")
s = MicroSpice.solve(nl, [:in, :gnd], [:out])
decibel(x) = 20 * log10(abs(x))
[decibel(only(s(f, [1, 0]))) for f in [1.4e6, 1.5e6, 1.62e6, 1.8e6]]
# output
4-element Vector{Float64}:
 12.883077832402897
 18.914298710446417
 27.65586909592151
 11.05634871505561

MicroSpice.Netlist — Type

Reads some input and parses that input as a netlist

MicroSpice.coil — Method

Inductance in nH of a coil with n turns d mm in diameter and l mm long

MicroSpice.inputImpedance — Method

Returns a function that takes a frequency and computes the impedance of a circuit from the specified input to the specified reference (ground) node.

MicroSpice.solve — Function

Returns a function that simulates a circuit with specified voltage inputs and outputs. That function takes as arguments the frequency and a list of input voltages and it returns a list of output voltages.

MicroSpice.solve — Method

Compute the voltages and injected currents for each node.

The input is a list of $(i, V_i)$ pairs which represent the imposed voltages and a list of injected currents as $(i, I_i)$. The nodes with imposed voltages should not overlap with the nodes with injected currents. All nodes without an imposed voltage are considered free-floating and their voltage level will be computed. Nodes with no injected current are considered isolated except for the connections specified in the connections in the circuit.

The result is a vector of node voltages and injected currents $[V_1 \ldots V_n, I_1 \ldots I_n]$

MicroSpice.xc — Method

Complex reactance of a capacitor at a given frequency

MicroSpice.xl — Method

Complex reactance of an inductor at a given frequency